% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{addEvents}
\alias{addEvents}
\title{addEvents}
\usage{
addEvents(node, ...)
}
\description{
addEvents(node) assigns all corresponding events of the contracts
in the respective leaf of the institution tree
}
