% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{addMarketObject2Contracts}
\alias{addMarketObject2Contracts}
\title{addMarketObject2Contracts}
\usage{
addMarketObject2Contracts(institution, yc, spread, cycle, ...)
}
\description{
addMarketObject2Contracts(institution, yc, cylce) assigns a given
MarketObject (YieldCurve) to the contracts which it applies to
}
