% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{assignContracts2Tree}
\alias{assignContracts2Tree}
\title{assignContracts2Tree}
\usage{
assignContracts2Tree(institution, ptf, ...)
}
\description{
assignContracts2Tree(institution, ptf) assigns a given Portfolio <ptf>
to the respective leaf of the institution tree while converting the contracts
of the given Portfolio <ptf> into a data.frame first.
}
