% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{assignEvents2Tree}
\alias{assignEvents2Tree}
\title{assignEvents2Tree}
\usage{
assignEvents2Tree(institution, riskFactors)
}
\description{
assignEvents2Tree(institution) assigns all corresponding events of the contracts
in the respective leaf of the institution tree
}
