% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{createInstitution}
\alias{createInstitution}
\title{createInstitution}
\usage{
createInstitution(name, ...)
}
\description{
createInstitution(name) function of class Institution 
creates the whole hierarchy of the institution with name <name>.
}
