% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCurve.R
\docType{methods}
\name{DefaultCurve}
\alias{DefaultCurve}
\alias{DefaultCurve-method}
\alias{DefaultCurve,ANY-method}
\title{\code{DefaultCurve}-class constructor}
\usage{
DefaultCurve(...)

\S4method{DefaultCurve}{ANY}(...)
}
\arguments{
\item{...}{}
}
\value{
An object of class \code{DefaultCurve} 
         containing the reference to the Java object
}
\description{
Create an instance of \code{DefaultCurve} class. The 
constructor will also create an instance of the respective
Java class in the running JVM.
}
\seealso{
\code{\link{YieldCurve}}
}
