% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscountingEngine.R
\docType{class}
\name{DiscountingEngine-class}
\alias{DiscountingEngine-class}
\title{A \code{\link{ValuationEngine}} implementing discounting of
future cash flows}
\description{
This class represents a \code{\link{ValuationEngine}} that
uses the future cash flow discounting method to compute a
value of a CT. The engine is implemented in Java so that
this class points to an instance of the respective Java
class.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\seealso{
\code{\link{DiscountingEngineModel, ValuationEngine}}
}
