% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Default.R, R/Institution.R
\docType{methods}
\name{default}
\alias{default}
\alias{generateDefaultContracts}
\alias{generateDefaultContracts,ContractType,list,character,data.frame-method}
\alias{default,Node,list,character,list-method}
\alias{determineDefault}
\title{Default}
\usage{
default(object, defaults, from, rawCtrs, ...)

generateDefaultContracts(object, defaults, from, ctr, ...)

\S4method{generateDefaultContracts}{ContractType,list,character,data.frame}(object, defaults, from, ctr)

\S4method{default}{Node,list,character,list}(object, defaults, from, rawCtrs)

determineDefault(node, defaults, from, rawCtrs)
}
\arguments{
\item{object}{object The \code{Institution}-object for which the contracts indicate a default risk}

\item{defaults}{A list representing the default curves for the different economic sectors}

\item{from}{A character defining a date for the default simulation}
}
\description{
Default
}
