% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReferenceIndex.R
\name{defineReferenceIndex}
\alias{defineReferenceIndex}
\title{defineReferenceIndex}
\usage{
defineReferenceIndex(rfdfp)
}
\arguments{
\item{rfdfp}{character  Pathname of csv data file for the referenceIndex}
}
\value{
list       list with S4 refs to class=ReferenceIndex objects
}
\description{
defineReferenceIndex(rfdfp) takes as input a risk-factor-data-filepath 
  reads this data and returns an initialized list of
  ReferenceIndex objects with risk factors from this csv file.
}
