% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Duration.R
\name{duration}
\alias{duration}
\title{\code{duration}}
\usage{
duration(x, type="macaulay", yield, price, isPercentage=TRUE, by)
}
\arguments{
\item{yield}{a \code{numeric}, an Object of class \code{YieldCurve} or 
\code{DynamicYieldCurve} that describes the spot rate term structure
or indicates the yield used to calculate the duration.}

\item{by}{a \code{character} indicating the date as for which the net present value
is calculated.}

\item{type}{a \code{character} defining the type of duration; possible types are 
'fisher-weil', 'macaulay' (default) and 'dollar'.}

\item{price}{a \code{numeric}, indicating the price used for calculating the yield to maturity
of each contract.}

\item{isPercentage}{a \code{numeric}, indicating if the 'yield' is passed as percentage 
(TRUE) or as fraction (FALSE).}

\item{x}{the contract or portfolio of contracts for which to calculate 
the duration.}
}
\value{
a \code{numeic} containing the calculated duration.
}
\description{
Function which calculates the duration of a portfolio or a contract.
}
\details{
For the Macaulay duration, if \code{yield} is not provided, \code{price} should
be provided and is used to calculate the  \code{yield}.
For the Fisher-Weil duration, \code{yieldCurve} must be specified. In this
case the argument \code{price} has no effect.
}
