% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Events.R, R/EventSeries.R
\docType{methods}
\name{events}
\alias{events}
\alias{events,FEMSContract,character,missing-method}
\alias{events,FEMSContract,AD0,missing-method}
\alias{events,FEMSContract,timeDate,missing-method}
\title{Derive the events for a \code{ContractType}}
\usage{
events(object, processor, riskFactors, ...)

\S4method{events}{FEMSContract,character,missing}(object, processor, riskFactors)

\S4method{events}{FEMSContract,AD0,missing}(object, processor, riskFactors)

\S4method{events}{FEMSContract,timeDate,missing}(object, processor, riskFactors)
}
\arguments{
\item{object}{The \code{ContractType} for which to derive the events}

\item{ad}{The analysis date as per which all future events are to be derived}

\item{model}{(optional) The \code{RiskFactorConnector} conditional to which events are computed}
}
\value{
A \code{EventSeries} object containing the resulting events
}
\description{
Method generating the \code{\link{EventSeries}} of a contract, cf. section Details.
}
\details{
The events of a contract are directly derived from the legal obligations 
defined therein. Events mark the exchange of a cash flow among the parties
involved in the contract or changes in the inner states of the contract
affecting cash flows to be exchanged in future. 
As such, \code{events} methods generates the
raw results that are the input for many financial analyses.
}
\seealso{
\link{ContractType}, \link{EventSeries}, \link{RiskFactorConnector}
}
