% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EventSeries.R
\docType{class}
\name{EventSeries-class}
\alias{EventSeries-class}
\title{S4 class EventSeries holds a dataframe with the cashflow events generated for
a single Actus Contract, along with the contractID, contractType, statusDate
contructors: EventSeries()  EventSeries(<Contract>,<rfs>) 2B exported}
\description{
S4 class EventSeries holds a dataframe with the cashflow events generated for
a single Actus Contract, along with the contractID, contractType, statusDate
contructors: EventSeries()  EventSeries(<Contract>,<rfs>) 2B exported
}

