% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEMSContract.R
\docType{class}
\name{FEMSContract-class}
\alias{FEMSContract-class}
\title{FEMSContract is  Reference Class parent to all FEMS Contract Types}
\description{
This class is only used in an abstract sense in that it is
not intended to create instances of it. It's use is to
serve as parent to various implementations of FEMS CTs (Operations) as
a means for designing method-inheritance.'
}

