% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{fSensitivity}
\alias{fSensitivity}
\title{specific function for computing sensitivity analytics on a data.tree structure 
of class Node}
\usage{
fSensitivity(node, ...)
}
\description{
This function computes analytics individually for the leafs of a tree
The analytics to be computed must be passed as first argument.
This function thus subsumes the function of all three specialized 
functions above (which are commented out)
}
