% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiskFactorConnector.R, R/DefaultCurve.R,
%   R/YieldCurve.R, R/DiscountingEngine.R, R/DynamicYieldCurve.R
\docType{methods}
\name{get,RiskFactorConnector,character-method}
\alias{get,RiskFactorConnector,character-method}
\alias{get-method}
\alias{get,DefaultCurve,character-method}
\alias{get,YieldCurve,character-method}
\alias{get,ForeignExchangeRate,character-method}
\alias{get,ReferenceIndex,character-method}
\alias{get,DiscountingEngine,character-method}
\alias{get,ContractType,character-method}
\alias{get,ValuationEngine,character-method}
\alias{get,jobjRef,character-method}
\alias{get,DynamicYieldCurve,character-method}
\title{Retrieve a \code{\link{RiskFactor}} object from the 
\code{\link{RiskFactorConnector}}}
\usage{
\S4method{get}{RiskFactorConnector,character}(object, what)

\S4method{get}{DefaultCurve,character}(object, what, ...)

\S4method{get}{YieldCurve,character}(object, what, ...)

\S4method{get}{DiscountingEngine,character}(object, what)

\S4method{get}{DynamicYieldCurve,character}(object, what, ...)
}
\arguments{
\item{object}{The \code{\link{RiskFactorConnector}} object}

\item{what}{A character (vector) containing key(s) to
elements in \code{object} which are to be returned}

\item{...}{}
}
\value{
an object of class \code{\link{RiskFactor}} with 
       'MarketObjectCode'=\code{what} parameter
}
\description{
Each element of the \code{what} parameter is expected to
point to an element in the \code{RiskFactorConnector} by
their 'keys'. All elements that can be found in
\code{RiskFactorConnector} are returned.
}
\examples{
yc <- YieldCurve() # create a YieldCurve object
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(MarketObjectCode = "YC_Prim",
    Nodes = list(ReferenceDate = "2015-01-01", 
    Tenors = tenors, Rates = rates)))

ind <- Index() # create a ReferenceIndex object
times <- c("2015-01-01", "2016-01-01", "2017-01-01", "2018-01-01",
           "2019-01-01")
values <- c(100, 110, 120, 130, 140)
set(ind, what=list(MarketObjectCode = "CHF_SMI",
    Data=list(Dates=times,Values=values)))
    
rf <- RFConn() # create a RiskFactorConnector object
add(rf,list(yc,ind)) # add all risk factors to the RiskFactorConnector
get(rf, "YC_Prim")

}
\seealso{
add, remove
}
