% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{getEvents}
\alias{getEvents}
\title{getEvents}
\usage{
getEvents(node, cid, ...)
}
\description{
getEvents(node, cid) gets all corresponding events of the contract
in the respective leaf of the institution tree
}
