% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{getLeafsAsDataFrames}
\alias{getLeafsAsDataFrames}
\title{getLeafsAsDataFrames}
\usage{
getLeafsAsDataFrames(institution, ...)
}
\description{
getLeafsAsDataFrames(institution) converts each leaf and it's contracts
to a data.frame and returns a list of data.frames.
}
