% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{getPortfolioAsDataFrame,Portfolio-method}
\alias{getPortfolioAsDataFrame,Portfolio-method}
\title{getPortfolioAsDataFrame(ptf)}
\usage{
\S4method{getPortfolioAsDataFrame}{Portfolio}(ptf)
}
\arguments{
\item{ptf}{S4 ref to class=Portfolio object with list of contracts}
}
\value{
A data.frame with reduced attributes for given Portfolio
}
\description{
Defines a generic method on S4 Class Portfolio to transform
  portfolio ptf given as input parameter to a data.frame
  with reduced attributes.
  
  Attributes in data.frame:
   - ContractID
   - ContractRole
   - StatusDate
   - Currency
   - NotionalPrincipal
   - NominalInterestRate
   - SpreadRate
   - ContractDealDate
   - InitialExchangeDate
   - MaturityDate
   - MarketObjectCode
}
