% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{getPortfolioAsDataFrame}
\alias{getPortfolioAsDataFrame}
\title{getPortfolioAsDataFrame <ptf>     Generic method definition}
\usage{
getPortfolioAsDataFrame(ptf)
}
\arguments{
\item{ptf}{S4 reference Class=Portfolio Portfolio with a list of contracts}
}
\value{
A data.frame with reduced attributes for given Portfolio
}
\description{
Defines a generic method on S4 Class Portfolio to transform
  portfolio ptf given as input parameter to a data.frame
  with reduced attributes.
}
