% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCurve.R
\name{getRatesAsSeries}
\alias{getRatesAsSeries}
\title{Interpolate the DefaultCurve and get the rates for specified date range
from until to}
\usage{
getRatesAsSeries(object, dates, ...)
}
\description{
Interpolate the DefaultCurve and get the rates for specified date range
from until to
}
