% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YieldCurve.R
\docType{class}
\name{YieldCurve-class}
\alias{YieldCurve-class}
\title{A Reference Class extending \code{\link{RiskFactor}} class
and representing a yield curve risk factor}
\description{
Yield curves define a class of market risk factors 
that in the ACTUS model may directly affect future cash 
flows arising from a financial instrument, e.g. Floating-
rate bonds, or are used for valuation of instruments, 
e.g. in discounting.
}
\section{Fields}{

\describe{
\item{\code{jref}}{A rJava java object reference}
}}


\examples{
yc <- YieldCurve()
tenors <- c("1W", "1M", "6M", "1Y", "2Y", "5Y")
rates <- c(0.001, 0.0015, 0.002, 0.01, 0.02, 0.03)
set(yc, what = list(MarketObjectCode = "YC_Prim",
  Nodes = list(ReferenceDate = "2015-01-01T00", Tenors = tenors, Rates = rates)))

get(ind, "MarketObjectCode")
get(yc, "ReferenceDate")
get(yc, "Tenors")
get(yc, "Rates")

}
\seealso{
\code{\link{RiskFactor, ReferenceIndex, ForeignExchangeRate}}
}
