% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{Institution}
\alias{Institution}
\title{Class that contains the whole model of an enterprise or institution.}
\description{
The class inherits from class \code{Node} in package \code{data.tree}
It contains a hierarchical model structure.
}
\section{Super class}{
\code{\link[data.tree:Node]{data.tree::Node}} -> \code{Institution}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Institution-clone}{\code{Institution$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="AddChild"><a href='../../data.tree/html/Node.html#method-Node-AddChild'><code>data.tree::Node$AddChild()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="AddChildNode"><a href='../../data.tree/html/Node.html#method-Node-AddChildNode'><code>data.tree::Node$AddChildNode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="AddSibling"><a href='../../data.tree/html/Node.html#method-Node-AddSibling'><code>data.tree::Node$AddSibling()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="AddSiblingNode"><a href='../../data.tree/html/Node.html#method-Node-AddSiblingNode'><code>data.tree::Node$AddSiblingNode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Climb"><a href='../../data.tree/html/Node.html#method-Node-Climb'><code>data.tree::Node$Climb()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Do"><a href='../../data.tree/html/Node.html#method-Node-Do'><code>data.tree::Node$Do()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Get"><a href='../../data.tree/html/Node.html#method-Node-Get'><code>data.tree::Node$Get()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Navigate"><a href='../../data.tree/html/Node.html#method-Node-Navigate'><code>data.tree::Node$Navigate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Prune"><a href='../../data.tree/html/Node.html#method-Node-Prune'><code>data.tree::Node$Prune()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="RemoveAttribute"><a href='../../data.tree/html/Node.html#method-Node-RemoveAttribute'><code>data.tree::Node$RemoveAttribute()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="RemoveChild"><a href='../../data.tree/html/Node.html#method-Node-RemoveChild'><code>data.tree::Node$RemoveChild()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Revert"><a href='../../data.tree/html/Node.html#method-Node-Revert'><code>data.tree::Node$Revert()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Set"><a href='../../data.tree/html/Node.html#method-Node-Set'><code>data.tree::Node$Set()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="Sort"><a href='../../data.tree/html/Node.html#method-Node-Sort'><code>data.tree::Node$Sort()</code></a></span></li>
<li><span class="pkg-link" data-pkg="data.tree" data-topic="Node" data-id="initialize"><a href='../../data.tree/html/Node.html#method-Node-initialize'><code>data.tree::Node$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Institution-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Institution-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Institution$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
