% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{mergePortfolios,Portfolio,Portfolio-method}
\alias{mergePortfolios,Portfolio,Portfolio-method}
\title{mergePortfolios(ptf1, ptf2)}
\usage{
\S4method{mergePortfolios}{Portfolio,Portfolio}(ptf1, ptf2)
}
\arguments{
\item{ptf1}{S4 ref to class=Portfolio object with list of contracts}

\item{ptf2}{S4 ref to class=Portfolio object with list of contracts}
}
\value{
S4 ref to class=Portfolio object with merged list of contracts
}
\description{
mergePortfolios(ptf1, ptf2) takes as input two S4 Class Portfolio to merge
  them into one Portfolio holding all the required of an institution.
}
