% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{mergePortfolios}
\alias{mergePortfolios}
\title{mergePortfolios <ptf1, ptf2>     Generic method definition}
\usage{
mergePortfolios(ptf1, ptf2)
}
\arguments{
\item{ptf1}{S4 ref to class=Portfolio object with list of contracts}

\item{ptf2}{S4 ref to class=Portfolio object with list of contracts}
}
\value{
S4 ref to class=Portfolio object with merged list of contracts
}
\description{
Defines a generic method on S4 Class Portfolio to merge
  two portfolios ptf1 and ptf2 given as input parameter into one Portfolio 
  holding all the required of an institution.
}
