% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperationalCF.R
\docType{class}
\name{OperationalCF-class}
\alias{OperationalCF-class}
\title{OperationalCF Contract class definition}
\usage{
OperationalCF(ContractID, pattern, args, ...)
}
\arguments{
\item{pattern}{A function evaluating the pattern of generated cash flows.}

\item{args}{The list of arguments used when evaluating the pattern}
}
\description{
An operational cash flows contract represents any operational activity 
in monetary units within an organization.
}

