% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Portfolio.R
\name{samplePortfolio}
\alias{samplePortfolio}
\title{samplePortFolio}
\usage{
samplePortfolio(cdfn, type)
}
\arguments{
\item{cdfn}{character string -  a contract-data-filepath}

\item{type}{character string - defines if operations or contracts}
}
\value{
Portfolio s4 object initialized with the data from the input files
}
\description{
samplePortfolio(cdfn) takes as input a contracts-data-filepath 
  reads this data and returns an initialized
  Portfolio object with contracts and from this csv file.
}
