% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sensitivity.R, R/Institution.R
\docType{methods}
\name{sensitivity}
\alias{sensitivity}
\alias{sensitivity,Node,YieldCurve-method}
\title{Derive the sensitivity of a contract's value with respect
to a change in risk factors}
\usage{
sensitivity(object, yield, ...)

\S4method{sensitivity}{Node,YieldCurve}(object, yield)
}
\arguments{
\item{object}{object \code{Node}}

\item{yield}{object \code{YieldCurve}}
}
\value{
A \code{numeric} object giving the sensitivity
}
\description{
Different sensitivity-concepts can be derived from a financial
instrument or its analytical value, respectively. Currentently, 
these are only the Fisher-Weil for Maturity-contracts.
}
\details{
Fisher-Weil Duration is a measure for the sensitivity of the net present
value of a Maturity instrument to a parallel Shift in the related
yield curve. This measure is based on the discounted cash flows of the 
instrument.
}
