% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeBuckets.R
\docType{methods}
\name{[,timeBuckets,numeric,missing-method}
\alias{[,timeBuckets,numeric,missing-method}
\title{Subscript method for class timeBuckets.
 
Since class timeBuckets provides a partition of a time interval, the value 
returned is the sequences of timeBuckets for the minimal enclosing set
of indices.}
\usage{
\S4method{[}{timeBuckets,numeric,missing}(x, i, j, ..., drop = TRUE)
}
\description{
Subscript method for class timeBuckets.
 
Since class timeBuckets provides a partition of a time interval, the value 
returned is the sequences of timeBuckets for the minimal enclosing set
of indices.
}
