% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{assignContracts2Tree}
\alias{assignContracts2Tree}
\title{Function of assign a given portfolio (from class Portfolio) to the respective
leaf of the institution tree.}
\usage{
assignContracts2Tree(institution, ptf, ...)
}
\description{
Function of assign a given portfolio (from class Portfolio) to the respective
leaf of the institution tree.
}
