% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Institution.R
\name{createInstitution}
\alias{createInstitution}
\title{Function of class Institution to create the whole hierarchy of the institution.}
\usage{
createInstitution(name, ...)
}
\description{
Function of class Institution to create the whole hierarchy of the institution.
}
